backend/requirements.txt
fastapi
uvicorn
python-multipart
jinja2

openai
python-dotenv
numpy
pandas
librosa
soundfile
noisereduce
moviepy
pydub
scipy
av